requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
Pillow>=8.2.0,<8.3.0
django-cors-headers>=3.7.0,<3.8.0
djoser>=2.1.0,<2.2.0
drf-spectacular>=0.15.1,<0.16
channels>=3.0.0,<4.0.0
channels-redis>=3.3.0,<4.0.0
redis>=4.2.0,<5.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
uvicorn[standard]>=0.20,<1.0
